# Optimise serialisation