Optimise serialisation by getting rid of loops and replacing them with chained generators.

Stats on loading / dumping 10 000 000 candidate pairs:
**Before:** Dump: 5.3s  Load: 25.2s
**Before:** Dump: 3.8s  Load: 11.0s

Relatedly, I find it really counterintuitive that loading takes longer than dumping. I suspect `array` may be copying data around too much.