@nbgl 2018-08-17
Also! It's possible to shave another 1.5s or so from the loading time by removing the try/except block from inside `_entries_iterable`. This lets us `return` the iterable instead of using `yield from`.

The exception still has to be caught so we can make it a `ValueError`, but we can wrap the lines that actually exhaust the iterator in try/except.

I chose not to do this, as I felt it would make the logic of the code too unclear.